Disable creation, update and removing items for example version
 Changes to be committed:
	modified:   app/settings.py
	modified:   organization/views/items.py

+++ organization/views/items.py
@@ -2,6 +2,7 @@ from django.http import HttpResponse
 from django.apps import apps
 from app.responses import error, ok
 from app.view import SecureView
+from app.settings import VERSION
 from organization.elements.elements import get_org_elms_public_info, get_org_elms_public_info_str, get_org_elms_private_info, FIELD_PARAM_TYPE
 from organization.views.utils.list_item_fields import get_items_list, get_item_list_section
 import json
@@ -52,6 +53,8 @@ class ItemsSectionView(ItemTypeView):
     
 class CreateItemView(ItemTypeView):
     def post(self, request):
+        if VERSION == 'example':
+            return ok()
         try:
             input_fields = get_input_field_values(self.item_type, request)
             created_item = self.item_model.objects.create(**input_fields)
@@ -86,6 +89,8 @@ class ItemView(ItemTypeView):
         return ok(item_fields=item_fields)
 class UpdateItemView(ItemTypeView):
     def patch(self, request):
+        if VERSION == 'example':
+            return ok()
         try:
             print('A')
             input_fields = get_input_field_values(self.item_type, request)
@@ -100,6 +105,8 @@ class UpdateItemView(ItemTypeView):
             return error(409, str(e))
 class DeleteItemView(ItemTypeView):
     def delete(self, request):
+        if VERSION == 'example':
+            return ok()
         try:
             self.item_model.objects.filter(id=request.GET.get("item_id")).delete()
             return ok()
